updated package name and changed rest endpoint

+++ src/main/java/com/leoantsmith/smart_living_backend/controller/dtos/CoordinatesDTO.java
@@ -1,22 +1,22 @@
 package com.leoantsmith.smart_living_backend.controller.dtos;
 
 public class CoordinatesDTO {
-    long lon;
-    long lat;
+    double lon;
+    double lat;
 
-    public long getLon() {
+    public double getLon() {
         return lon;
     }
 
-    public void setLon(long lon) {
+    public void setLon(double lon) {
         this.lon = lon;
     }
 
-    public long getLat() {
+    public double getLat() {
         return lat;
     }
 
-    public void setLat(long lat) {
+    public void setLat(double lat) {
         this.lat = lat;
     }
 }
